Create config.py file for global config variables
The jenkins configuration variables were not being properly initialized
in the Config class of the utils module, so I moved them to their own
config.py class that will sit in the parent directory of the src module
to prevent circular dependencies. Eventually we can move all of the
configuration settings to this module and keep our utils and our config
separate.